+++ docs/whatsnew/changelog.rst
@@ -20,7 +20,7 @@ New features
 ~~~~~~~~~~~~
 .. Add here new public features (do not delete this comment)
 
-* `PLS` (Partial Least Squares regression) method added.
+* `PLSRegression` (Partial Least Squares regression) method added.
 
 * `read` method now handle any url pointing
   to a spectrochempy readeable file. An url to a compressed (zip) files are also accepted.